refactor to adhere to architectural constraints
